requirements.txt
discord.py
python-dotenv~=1.1.0
pillow~=11.2.1
flask